replaced all float data types with double

+++ includes/canvas.h
@@ -6,7 +6,7 @@
 /*   By: bpetrovi <bpetrovi@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/05 22:04:04 by bpetrovi          #+#    #+#             */
-/*   Updated: 2026/05/09 14:42:21 by bpetrovi         ###   ########.fr       */
+/*   Updated: 2026/05/11 21:41:22 by bpetrovi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ typedef struct s_canvas
 	int		height;
 }	t_canvas;
 
-t_tuple	color(float r, float g, float b);
+t_tuple	color(double r, double g, double b);
 int		init_canvas(t_canvas *canvas, int x, int y);
 void	canvas_to_ppm(t_canvas *canvas);
 void	write_pixel(t_canvas *canvas, int x, int y, t_tuple color);
